Process data in chunks greater than one line.
When testing I realised that DynamicVectorRangeLayer was significatly slower than
VectorRangeLayer, by a couple of orders of magnitude. The former was rendering
the vector layer one line at a time, to save memory, whereas the later renders
the entire layer into memory, which has been problematic in the past.

On a test case, with a range of 24000x25000, I saw the dynamic version take
247 seconds in RasterizeLayer, whereas the render all at once took just 0.8
seconds (though the former uses at most 200KB of memory, whereas the later
takes 4.4GB!).

I did a little playing with chunk sizes, which we'll want to do for CUDA support
anyway, and I got almost identical time performance in the dynamic version as the
full raster version, whilst still using only 100MB of memory (with a 512 line
chunk). It did spend a little more time in RasterizeLayer, but I think we spent
less time in _eval and friends in operations.py due to the larger chunk size and
that offset the difference.