Apply distance based rouding when calculating layer pixel sizes.
Before this change we always called math.ceil to round up pixel
dimensions when going from degress/pixelscale, the theory being that
it is better to round up than lose data.

However, we saw in practice that some small errors below the floating
point epsilon value for the machine, when applied across large images,
could cause ceil to sometimes round up when in fact the error was still
only in the microns range.

Now we check the value and round up only if the error is 1 meter or more,
given that even at 400k pixels wide on jung data that's 100 meter per pixel.